% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client.R
\name{api_call}
\alias{api_call}
\title{Make API call}
\usage{
api_call(model, func, req, host = "http://127.0.0.1:5000")
}
\arguments{
\item{model}{model name}

\item{func}{function name}

\item{req}{request data, typically a list or data}

\item{host}{the host of the API server}
}
\description{
Make API call
}
\examples{
\dontrun{
api_run("iris", "pred", iris)
}

}

