% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client.R
\name{api_deploy}
\alias{api_deploy}
\title{Deploy a model to API}
\usage{
api_deploy(..., model_name, host = "http://127.0.0.1:5000")
}
\arguments{
\item{...}{list of objects}

\item{host}{API server}

\item{.model_name}{the name of the model}
}
\description{
Deploy a model to API
}

