% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client.R
\name{api_expose}
\alias{api_expose}
\title{Make a function to available for API call}
\usage{
api_expose(f)
}
\arguments{
\item{f}{the function to expose}
}
\description{
By default, a function should not be exposed to API for security concern.
We should only expose the functions we know that has no side effect and
necessary for user to use.
}

