% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.R
\name{json_wrapper}
\alias{json_wrapper}
\title{Wrap function output to be compatible with jsonify in Flask}
\usage{
json_wrapper(x)
}
\arguments{
\item{x}{R function output object}
}
\description{
The output object from function should be a names list, otherwise
the jsonify function in Flask would fail to convert. This function will
check if the object is a named list, if not it will wrap the whole object
into an element of a list with name response: list(response = object)
}

