% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.R
\name{rds_to_env}
\alias{rds_to_env}
\title{Load all RDS files from a folder into global enrironment}
\usage{
rds_to_env(folder)
}
\arguments{
\item{folder}{folder containing the RDS objects}
}
\description{
Load all RDS files from a folder into global enrironment
}
\examples{
folder <- "rdsfiles"
rds_to_env(folder)
}

